네이밍 컨벤션 변경 및 코드 중복되는거 통일
