+++ src/main/java/com/sw/journal/journalcrawlerpublisher/controller/RecommendedArticleController.java
@@ -1,12 +1,8 @@
 package com.sw.journal.journalcrawlerpublisher.controller;
 
-import com.sw.journal.journalcrawlerpublisher.domain.Category;
 import com.sw.journal.journalcrawlerpublisher.domain.Member;
 import com.sw.journal.journalcrawlerpublisher.domain.OurArticle;
-import com.sw.journal.journalcrawlerpublisher.domain.Tag;
-import com.sw.journal.journalcrawlerpublisher.service.SungbinTestService;
-import lombok.Getter;
-import lombok.Setter;
+import com.sw.journal.journalcrawlerpublisher.service.RecommendArticleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,13 +10,13 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/test")
-public class SungbinTestController {
+public class RecommendedArticleController {
 
-    private final SungbinTestService sungbinTestService;
+    private final RecommendArticleService recommendArticleService;
 
     @Autowired
-    public SungbinTestController(SungbinTestService sungbinTestService) {
-        this.sungbinTestService = sungbinTestService;
+    public RecommendedArticleController(RecommendArticleService recommendArticleService) {
+        this.recommendArticleService = recommendArticleService;
     }
 
     // 유저 선호 카테고리 기사 검색
@@ -28,6 +24,6 @@ public class SungbinTestController {
     public List<OurArticle> getArticlesByUserFavoriteCategories(@PathVariable Long userId) {
         Member member = new Member();
         member.setId(userId);
-        return sungbinTestService.findByUserFavoriteCategories(member);
+        return recommendArticleService.findByUserFavoriteCategories(member);
     }
 }
