+++ src/main/java/com/sw/journal/journalcrawlerpublisher/service/OurArticleService.java
@@ -1,22 +1,22 @@
 package com.sw.journal.journalcrawlerpublisher.service;
 
-import com.sw.journal.journalcrawlerpublisher.domain.*;
+import com.sw.journal.journalcrawlerpublisher.domain.Category;
+import com.sw.journal.journalcrawlerpublisher.domain.OurArticle;
+import com.sw.journal.journalcrawlerpublisher.domain.Tag;
 import com.sw.journal.journalcrawlerpublisher.repository.CategoryRepository;
 import com.sw.journal.journalcrawlerpublisher.repository.OurArticleRepository;
 import com.sw.journal.journalcrawlerpublisher.repository.TagRepository;
-import com.sw.journal.journalcrawlerpublisher.repository.UserFavoriteCategoryRepository;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-// wild-mantle 2024-07-08
-@Getter @Setter
+@Getter
+@Setter
 @Service
-public class SungbinTestService {
+public class OurArticleService {
 
     @Autowired
     private OurArticleRepository ourArticleRepository;
@@ -28,10 +28,7 @@ public class SungbinTestService {
     private TagRepository tagRepository;
 
     @Autowired
-    private UserFavoriteCategoryRepository userFavoriteCategoryRepository;
-
-    @Autowired
-    public SungbinTestService(OurArticleRepository ourArticleRepository) {
+    public OurArticleService(OurArticleRepository ourArticleRepository) {
         this.ourArticleRepository = ourArticleRepository;
     }
 
@@ -73,14 +70,5 @@ public class SungbinTestService {
     // n개 카테고리, n개 태그로 검색
     public List<OurArticle> findByCategoriesAndTags(List<Category> categories, List<Tag> tags) {
         return ourArticleRepository.findByCategoriesAndTags(categories, tags, tags.size());
-    }
-
-    // 유저가 선호하는 카테고리의 기사 검색
-    public List<OurArticle> findByUserFavoriteCategories(Member member) {
-        List<UserFavoriteCategory> favoriteCategories = userFavoriteCategoryRepository.findByMember(member);
-        List<Category> categories = favoriteCategories.stream()
-                .map(UserFavoriteCategory::getCategory)
-                .collect(Collectors.toList());
-        return findByCategories(categories);
     }
 }
